Move dashboard specs next to source files
