+++ dashboard/src/components/InstitutionLogo.spec.ts
