+++ dashboard/src/helpers/logs.spec.ts
