+++ dashboard/src/stores/auth.spec.ts
